+++ src/components/Main/ApartMap/MapLoading.tsx
@@ -1,3 +1,5 @@
+// 네이버 지도 로딩 컴포넌트
+
 import styles from "./MapLoading.module.scss";
 
 export default function MapLoading() {
